requirements.txt
mysql-connector-python
requests
flake8